Fix crate version and addition
Bumping the addition introduced some linting errors which have been
fixed as part of this commit.

We're opting to use git tags and github releases to mark versions. If
the crate is to be published, the CI job should populate the version
before publishing based on the git tag triggering the CI.